# doc/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NexusHub V1 API",
    "version": "1.0.0-design",
    "description": "开发前契约，尚无运行服务。全部业务POST，只有CSRF引导为GET。跨字段领域规则见模块文档。"
  },
  "servers": [
    {
      "url": "/",
      "description": "同源部署"
    }
  ],
...
  "paths": {
...
    "/api/games/create": {
      "post": {
        "operationId": "createGame",
        "tags": [
          "game"
        ],
        "summary": "创建对局",
        "description": "权限：game:play + 版本可读且未归档。",
        "security": [
          {
            "SessionCookie": [],
            "CsrfHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功（异步创建仅表示已入队）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResponse"
                }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "权限或CSRF错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "不存在或不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "版本、状态或幂等冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "规则校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "限流或队列已满",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务异常",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GameCreateRequest"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 64,
              "minLength": 16,
              "pattern": "^[A-Za-z0-9_-]+$"
            },
            "description": "同一用户、同一路径、相同业务动作重试复用，建议UUID"
          }
        ]
      }
    },
    "/api/games/detail": {
      "post": {
        "operationId": "gameDetail",
        "tags": [
          "game"
        ],
        "summary": "对局详情",
        "description": "权限：game:play + 所有者。",
        "security": [
          {
            "SessionCookie": [],
            "CsrfHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功（异步创建仅表示已入队）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResponse"
                }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "权限或CSRF错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "不存在或不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "版本、状态或幂等冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "规则校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "限流或队列已满",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务异常",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdRequest"
              }
            }
          }
        }
      }
    },
    "/api/games/actions/play": {
      "post": {
        "operationId": "playMove",
        "tags": [
          "game"
        ],
        "summary": "出牌",
        "description": "权限：game:play + 所有者 + 当前人类座位。",
        "security": [
          {
            "SessionCookie": [],
            "CsrfHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功（异步创建仅表示已入队）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResponse"
                }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "权限或CSRF错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "不存在或不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "版本、状态或幂等冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "规则校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "限流或队列已满",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务异常",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayRequest"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 64,
              "minLength": 16,
              "pattern": "^[A-Za-z0-9_-]+$"
            },
            "description": "同一用户、同一路径、相同业务动作重试复用，建议UUID"
          }
        ]
      }
    },
    "/api/games/actions/pass": {
      "post": {
        "operationId": "passMove",
        "tags": [
          "game"
        ],
        "summary": "不出",
        "description": "权限：game:play + 所有者 + 当前人类座位。",
        "security": [
          {
            "SessionCookie": [],
            "CsrfHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功（异步创建仅表示已入队）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResponse"
                }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "权限或CSRF错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "不存在或不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "版本、状态或幂等冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "规则校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "限流或队列已满",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务异常",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GameVersionRequest"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 64,
              "minLength": 16,
              "pattern": "^[A-Za-z0-9_-]+$"
            },
            "description": "同一用户、同一路径、相同业务动作重试复用，建议UUID"
          }
        ]
      }
    },
    "/api/games/legal-moves/search": {
      "post": {
        "operationId": "legalMoves",
        "tags": [
          "game"
        ],
        "summary": "合法行动分页",
        "description": "权限：game:play + 所有者。",
        "security": [
          {
            "SessionCookie": [],
            "CsrfHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功（异步创建仅表示已入队）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegalMovePageResponse"
                }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "权限或CSRF错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "不存在或不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "版本、状态或幂等冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "规则校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "限流或队列已满",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务异常",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LegalMovesRequest"
              }
            }
          }
        }
      }
    },
    "/api/games/hint": {
      "post": {
        "operationId": "hint",
        "tags": [
          "game"
        ],
        "summary": "同步提示分析",
        "description": "权限：game:play + 所有者 + 当前人类座位。V1 同步返回一条规则合法建议，proofStatus 固定 UNKNOWN；后续可扩展为异步 SolveTask。",
        "security": [
          {
            "SessionCookie": [],
            "CsrfHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功（同步返回）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HintResponseResponse"
                }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "权限或CSRF错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "不存在或不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "版本、状态或幂等冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "规则校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "限流或队列已满",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务异常",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GameVersionRequest"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 64,
              "minLength": 16,
              "pattern": "^[A-Za-z0-9_-]+$"
            },
            "description": "同一用户、同一路径、相同业务动作重试复用，建议UUID"
          }
        ]
      }
    },
    "/api/games/abandon": {
      "post": {
        "operationId": "abandonGame",
        "tags": [
          "game"
        ],
        "summary": "放弃",
        "description": "权限：game:play + 所有者。",
        "security": [
          {
            "SessionCookie": [],
            "CsrfHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功（异步创建仅表示已入队）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResponse"
                }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "权限或CSRF错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "不存在或不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "版本、状态或幂等冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "规则校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "限流或队列已满",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务异常",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GameVersionRequest"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 64,
              "minLength": 16,
              "pattern": "^[A-Za-z0-9_-]+$"
            },
            "description": "同一用户、同一路径、相同业务动作重试复用，建议UUID"
          }
        ]
      }
    },
    "/api/games/restart": {
      "post": {
        "operationId": "restartGame",
        "tags": [
          "game"
        ],
        "summary": "新建重开对局",
        "description": "权限：game:play + 所有者。",
        "security": [
          {
            "SessionCookie": [],
            "CsrfHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功（异步创建仅表示已入队）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResponse"
                }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "权限或CSRF错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "不存在或不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "版本、状态或幂等冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "规则校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "限流或队列已满",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务异常",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GameRestartRequest"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 64,
              "minLength": 16,
              "pattern": "^[A-Za-z0-9_-]+$"
            },
            "description": "同一用户、同一路径、相同业务动作重试复用，建议UUID"
          }
        ]
      }
    },
...
    "/api/games/replay/detail": {
      "post": {
        "operationId": "gameReplay",
        "tags": [
          "game"
        ],
        "summary": "复盘分页",
        "description": "权限：game:play + 所有者。",
        "security": [
          {
            "SessionCookie": [],
            "CsrfHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功（异步创建仅表示已入队）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplayResponse"
                }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "权限或CSRF错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "不存在或不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "版本、状态或幂等冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "规则校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "限流或队列已满",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务异常",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplayRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "NEXUSHUB_SESSION"
      },
      "CsrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-TOKEN"
      }
    },
    "schemas": {
...
      "IdRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 19,
            "pattern": "^[1-9][0-9]{0,18}$",
            "description": "正整数 BIGINT 的十进制字符串，服务端还校验 <= 9223372036854775807"
          }
        },
        "required": [
          "id"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "integer",
            "minimum": 40000,
            "maximum": 59999
          },
          "message": {
            "type": "string",
            "maxLength": 500
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorData"
              },
              {
                "type": "null"
              }
            ]
          },
          "requestId": {
            "type": "string",
            "maxLength": 64
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "requestId"
        ]
      },
...
      "GameResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "const": 200,
            "type": "integer"
          },
          "message": {
            "type": "string",
            "maxLength": 500
          },
          "data": {
            "$ref": "#/components/schemas/Game"
          },
          "requestId": {
            "type": "string",
            "maxLength": 64
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "requestId"
        ]
      },
...
    }
  }
}